requirements.txt
Flask==2.3.3
pandas==2.0.3
python-dotenv==1.0.0
google-generativeai==0.5.4
folium==0.14.0
requests==2.31.0